requirements.txt
Flask==3.1.1
tensorflow==2.18.0
numpy==2.0.2
pillow==10.1.0
gunicorn==21.2.0
